User mapping (#26)
* Added strategies and partially companies

* Added strategies and partially companies

* Finish with strategies view

* Prepared strategy, company, ta

* Switch all to user_id

* Fix black and flake8

* Added cron job for run

* Delete .idea directory

* Updated UI to use token

* Use authorization for every endpoint

+++ backend/app/web/api/company/views.py
@@ -8,6 +8,7 @@ from app.web.api.company.scheme import (
     CompanyModelInputDTO,
     CompanyModelPatchDTO,
 )
+from app.web.deps import CurrentUser, check_owner_or_superuser
 from fastapi import APIRouter, Depends
 
 router = APIRouter()
@@ -17,34 +18,40 @@ logger = logging.getLogger(__name__)
 @router.get("/{company_id}", response_model=CompanyModelDTO)
 async def get_company_model(
     company_id: int,
+    current_user: CurrentUser,
     company_dao: CompanyDAO = Depends(),
 ) -> CompanyModel:
-    return await company_dao.get_company_model(company_id)
+    exist_company = await company_dao.get_company_model(company_id)
+    await check_owner_or_superuser(exist_company.user_id, current_user)
+    return exist_company
 
 
 @router.get("/", response_model=List[CompanyModelDTO])
 async def get_company_models(
+    current_user: CurrentUser,
     limit: int = 100,
     offset: int = 0,
     company_dao: CompanyDAO = Depends(),
 ) -> List[CompanyModel]:
-    return await company_dao.get_all_companies(limit=limit, offset=offset)
+    return await company_dao.get_all_companies(
+        limit=limit,
+        offset=offset,
+        user_id=current_user.id,
+    )
 
 
 @router.post("/")
 async def create_company_model(
     new_company_object: CompanyModelInputDTO,
+    current_user: CurrentUser,
     company_dao: CompanyDAO = Depends(),
 ) -> None:
     company = await company_dao.create_company_model(
         tiker=new_company_object.tiker,
-        name=new_company_object.name
-        if new_company_object.name
-        else new_company_object.tiker,
+        name=new_company_object.name,
         company_type=new_company_object.type,
-        strategies=[strategy.id for strategy in new_company_object.strategies]
-        if new_company_object.strategies
-        else None,
+        strategies=new_company_object.strategies,
+        user_id=current_user.id,
     )
     logger.info(f"Created company model for name={company.name}")
 
@@ -52,6 +59,7 @@ async def create_company_model(
 @router.post("/batch")
 async def create_company_batch_models(
     new_company_list: List[CompanyModelInputDTO],
+    current_user: CurrentUser,
     company_dao: CompanyDAO = Depends(),
 ) -> None:
     await company_dao.create_companies_models(
@@ -60,6 +68,7 @@ async def create_company_batch_models(
                 tiker=comp.tiker,
                 name=comp.name if comp.name else comp.tiker,
                 type=comp.type,
+                user=current_user,
             )
             for comp in new_company_list
         ],
@@ -70,8 +79,12 @@ async def create_company_batch_models(
 async def partial_update_company_model(
     company_id: int,
     updated_company: CompanyModelPatchDTO,
+    current_user: CurrentUser,
     company_dao: CompanyDAO = Depends(),
 ) -> None:
+    exist_company = await company_dao.get_company_model(company_id)
+    await check_owner_or_superuser(exist_company.user_id, current_user)
+
     await company_dao.update_company_model(
         company_id,
         updated_company.model_dump(),
@@ -83,14 +96,25 @@ async def partial_update_company_model(
 async def update_company_model(
     company_id: int,
     updated_company: CompanyModelInputDTO,
+    current_user: CurrentUser,
     company_dao: CompanyDAO = Depends(),
 ) -> None:
-    await company_dao.update_company_model(company_id, updated_company.model_dump())
+    exist_company = await company_dao.get_company_model(company_id)
+    await check_owner_or_superuser(exist_company.user_id, current_user)
+
+    await company_dao.update_company_model(
+        company_id,
+        updated_company.model_dump(),
+    )
 
 
 @router.delete("/{company_id}", status_code=204)
 async def delete_company_model(
     company_id: int,
+    current_user: CurrentUser,
     company_dao: CompanyDAO = Depends(),
 ) -> None:
+    exist_company = await company_dao.get_company_model(company_id)
+    await check_owner_or_superuser(exist_company.user_id, current_user)
+
     await company_dao.delete_company_model(company_id)
